Diagramme aktualisiert von sreitner
Merfach ise_ids, collect zu festen Uhrzeiten, collect Tages min/max-Wert; precision ; only_changed; line-color;....

+++ diagramm_collect.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 
 // Eintrag in Crontab
 //
@@ -7,6 +7,9 @@
 //
 // mit PHP-CLI
 // */1 * * * * /usr/bin/php -f /pfad-zu-homehub/diagramm_collect.php >/dev/null 2>&1
+
+ini_set('display_errors', 'on');
+ini_set('display_startup_errors', 'on');
 
 require_once(__DIR__.'/interface.php');
 
@@ -17,6 +20,10 @@ if (php_sapi_name() != "cli") echo '<pre>'.PHP_EOL;
 
 $minute = date('H')*60 + date('i');
 echo "Es ist ".$tage[date('w')].' '.date('d.m.Y H:i:s').', die '.$minute.'. Minute des Tages'.PHP_EOL;
+
+if (isset($_GET['dryrun']) or (!empty($argv[1]) and $argv[1]=='dryrun')) { echo '--- SIMULATION ---'.PHP_EOL; $_dryrun=true; $_verbose=true; }
+elseif (isset($_GET['test']) or (!empty($argv[1]) and $argv[1]=='test')) { echo '--- TEST ---'.PHP_EOL; $_test=true; $_verbose=true; }
+elseif (isset($_GET['verbose']) or (!empty($argv[1]) and $argv[1]=='verbose')) { $_verbose=true; }
 
 function read_config($file) {
 	if (!is_file($file)) return false;
@@ -39,23 +46,36 @@ function read_config($file) {
 // optinale Werte aus custom.json bzw. Standardwerte an Array übergeben
 function add_diagramm($custom) {
 	return array(
+		'type' => trim(strtolower($custom['component'])),
 		'only_changed' => ( !empty($custom['only_changed']) or strtolower($custom['component'])=="diagramm_change" ),
-		'precision' => ( empty($custom['precision']) ? 1 : intval($custom['precision']) ),
+		'precision' => ( !isset($custom['precision']) ? 1 : min(abs(intval($custom['precision'])), 5) ),
 	);
 }
 
 // Lese aus custom.json die ise_id, die geloggt werden sollen
 $data = read_config(__DIR__.'/config/custom.json');
-$json = json_decode($data, true);
+if (!$json = json_decode($data, true)) {
+	echo '!! Syntaxfehler custom.json !!'.PHP_EOL.PHP_EOL;
+	if (php_sapi_name() != "cli") echo '</pre>'.PHP_EOL;
+	exit(0);
+}
 
 foreach ($json['custom'] as $customs) {
 	foreach ($customs as $custom) {
 		if (!empty($custom['component'])) {
 			// <component> ist eines von ...
-			if (in_array(strtolower($custom['component']), array("diagramm", "diagramm_eckig", "mdiagramm", "diagramm_change"))) {
+			if ((in_array(trim(strtolower($custom['component'])), array("diagramm", "diagramm_eckig", "mdiagramm", "diagramm_change")) and empty($_test)) or ((trim(strtolower($custom['component'])) == 'diagramm_test') and !empty($_test))) {
 
-				// history auf ganzzahlige Werte zwischen 1 und 5000 begrenzen, Standard 200
+				if (preg_match('/\D/', $custom['ise_id'])) {
+					if (!empty($_verbose)) echo 'v  formatiere "'.$custom['ise_id'].'" zu "';
+					$custom['ise_id'] = trim(preg_replace('/\D+/', '-', $custom['ise_id']), '-');
+					if (!empty($_verbose)) echo $custom['ise_id'].'"'.PHP_EOL;
+				}
+
+				// <historyY auf ganzzahlige Werte zwischen 1 und 5000 begrenzen, Standard 200
 				$history = ( empty($custom['history']) ? 200 : max(1, min(intval($custom['history']), 5000)) );
+
+				if (isset($custom['collect'])) $custom['collect'] = trim($custom['collect']);
 
 				if (empty($custom['collect'])) {
 				// <collect> nicht definiert => immer sammeln
@@ -95,6 +115,12 @@ foreach ($json['custom'] as $customs) {
 					}
 					if (!isset($diagramm[$custom['ise_id']][$custom['collect']][$history])) echo '- überspringe '.$custom['ise_id'].' '.$history.', fällig um '.$custom['collect'].PHP_EOL;
 				}
+				elseif (preg_match('/(min|max)/i', $custom['collect'])) {
+				// Tagesniederst- und/oder -höchstwert
+					$custom['collect'] = preg_replace('/[\W]/', '-', $custom['collect']);
+					$diagramm[$custom['ise_id']][$custom['collect']][$history] = add_diagramm($custom);
+					echo '- hole '.$custom['ise_id'].' für Ermittlung '.$custom['collect'].'-Wert'.PHP_EOL;
+				}
 				else {
 					echo '- '.$custom['collect'].' für '.$custom['ise_id'].' '.$history.' kann nicht interpretiert werden'.PHP_EOL;
 				}
@@ -104,18 +130,27 @@ foreach ($json['custom'] as $customs) {
 	}
 }
 
+$datapoints = array_keys($diagramm);
+if (!empty($_verbose)) echo 'v  gefundene Indizes "'.implode('", "', $datapoints).'"'.PHP_EOL;
+
 // Mehrere ise_id einer Definition vereinzeln, damit alle Werte bei der CCU abgefragt werden
 // Erklärung: Wenn Datenpunkte als Array an die api_state in der interface.php übergeben, trennt diese nicht mehr nach Trennzeichen.
-$datapoints = array_keys($diagramm);
 $combine = array();
-foreach ($datapoints as $key => $ise_id) {
-	if (preg_match('/\D/', $ise_id)) {
-		$split = preg_split('/\D+/', $ise_id);
+$a_datapoints = array();
+foreach ($datapoints as $ise_id) {
+	if (strpos($ise_id, '-')) {
+		$split = explode('-', $ise_id);
+		if (!empty($_verbose)) echo 'v  trenne '.$ise_id.', '.count($split).' Datenpunkte: '.implode(' ', $split).PHP_EOL;
 		$combine[$ise_id] = $split;		// damit nachher wieder zusammengesetzt werden kann
-		array_splice($datapoints, $key, 1);		// diesen Eintrag löschen, dafür ...
-		$datapoints = array_merge($datapoints, $split);		// ... einzelne Datenpunkte anhängen
+		$a_datapoints = array_merge($a_datapoints, $split);
+	} else {
+		$a_datapoints[] = $ise_id;
 	}
 }
+
+$datapoints = array_unique($a_datapoints);
+unset($a_datapoints);
+if (!empty($_verbose)) echo 'v  abzufragende Datenpunkte '.implode(', ', $datapoints).PHP_EOL;
 
 if (!count($datapoints)) {
 	echo 'Beende, da keine relevanten Datenpunkte.'.PHP_EOL.PHP_EOL;
@@ -129,19 +164,44 @@ if (!is_object($xml->datapoint)) die('Fehler beim Auslesen der CCU-Abfrage!'.PHP
 foreach ($xml->datapoint as $datapoint) {
 	$ise_id = strval($datapoint->attributes()['ise_id']);
 	$value = strval($datapoint->attributes()['value']);
-	echo '- Wert: '.$ise_id.' = '.$value.PHP_EOL;
 	$values[$ise_id] = $value;
+	echo '- Wert: '.strval($ise_id).' = '.strval($value).PHP_EOL;
 }
 
 // Werte für Mehrfach-Diagramme zusammensetzen
 if (count($combine)) {
 	foreach ($combine as $ise_id => $multiple) {
+		if (!empty($_verbose)) echo 'v  setze '.$ise_id.' zusammen'.PHP_EOL;
 		$join = array();
 		foreach ($multiple as $datapoint) {
+			if (!isset($values[$datapoint])) {
+				echo '! Datenpunkt '.$datapoint.' in CCU-Antwort nicht gefunden'.PHP_EOL;
+				continue;
+			}
 			$join[] = $values[$datapoint];
 		}
 		$values[$ise_id] = $join;
 	}
+}
+
+function precision($values, $precision = 1) {
+// Numerische Werte in Dezimalzahl mit <precision> Nachkommastellen formatieren
+	global $_verbose;
+
+	if (is_array($values)) {
+		foreach ($values as $key => $value) {
+			if (!empty($_verbose)) echo 'v  formatiere '.$values[$key].' in ';
+			if (is_numeric($value)) $values[$key] = ( $precision ? number_format(floatval($values[$key]), $precision, '.', '') : intval($values[$key]) );
+			if (!empty($_verbose)) echo $values[$key].PHP_EOL;
+		}
+	}
+	elseif (is_numeric($values)) {
+		if (!empty($_verbose)) echo 'v  formatiere '.$values.' in ';
+		$values = ( $precision ? number_format(floatval($values), $precision, '.', '') : intval($values) );
+		if (!empty($_verbose)) echo $values.PHP_EOL;
+	}
+
+	return $values;
 }
 
 foreach ($diagramm as $ise_id => $collects) {
@@ -151,20 +211,20 @@ foreach ($diagramm as $ise_id => $collects) {
 		foreach ($collects as $collect => $histories) {
 			foreach ($histories as $history => $arr) {
 			// Einzelne Dateie je <collect> und <history> schreiben
+				if (!empty($_verbose)) echo 'v  verarbeite '.$ise_id.', collect '.$collect.', history '.$history.PHP_EOL;
 
-				$cfile = __DIR__.'/cache/diagramm_'.preg_replace('/\D/', '-', $ise_id).'_'.preg_replace('/\D/', '-', $collect).'_'.$history.'.csv';
+				$write[$ise_id] = precision($values[$ise_id], $arr['precision']);
+				$write[$ise_id] = ( is_array($write[$ise_id]) ? implode(';', $write[$ise_id]) : $write[$ise_id] );
+				if (!empty($_verbose)) echo 'v  CCU Werte '.strval($write[$ise_id]).PHP_EOL;
 
-				// Numerische Werte in Dezimalzahl mit <precision> Nachkommastellen formatieren
-				if (is_numeric($values[$ise_id])) $values[$ise_id] = number_format(floatval($values[$ise_id]), $arr['precision'], '.', '');
+				// Prefix für Beschriftung X-Achse
+				if (preg_match('/^\d+:\d+/', $collect)) $prefix = $tage[date('w')].' '.date('d.m.');
+				elseif (preg_match('/(min|max)/i', $collect)) $prefix = $tage[date('w')].' '.date('d.m.');
+				else $prefix = $tage[date('w')].' '.date('H:i');
+				# todo: Tag nur schreiben, wenn neuer Tag seit letztem Wert. Sollte mit $last[0] machbar sein #
+				if (!empty($_verbose)) echo 'v  Datensatz-Prefix '.($prefix).PHP_EOL;
 
-				// Werte für Mehrfach-Diagramme zusammensetzen
-				elseif (is_array($values[$ise_id])) {
-					$m_values = $values[$ise_id];
-					foreach ($m_values as $key => $value) {
-						if (is_numeric($value)) $m_values[$key] = number_format(floatval($m_values[$key]), $arr['precision'], '.', '');
-					}
-					$values[$ise_id] = implode(';', $m_values);
-				}
+				$cfile = __DIR__.'/cache/diagramm_'.( $arr['type']=='diagramm_test' ? 'test_' : '' ).$ise_id.'_'.preg_replace('/\W/', '-', $collect).'_'.$history.'.csv';
 
 				if (file_exists($cfile)) {
 				// cache Datei ist vorhanden
@@ -175,14 +235,70 @@ foreach ($diagramm as $ise_id => $collects) {
 					// Array auf maximal <history> Werte kürzen
 					while (count($csv) >= $history) array_shift($csv);
 
-					// Letzten gespeicherten Wert auslesen, überspringen falls unverändert
-					if (!empty($arr['only_changed'])) {
+					if (preg_match_all('/(min|max)/i', $collect, $minmax, PREG_PATTERN_ORDER)) {
+					// letzten Wert für Min/Max auslesen, Element überschreiben
 						if (count($csv)) {
+							$last = explode(';', rtrim($csv[count($csv)-1], ';'));
+							if ($last[0] == $prefix) {
+							// gleicher Tag wie letzter Wert => überschreiben
+								array_shift($last);
+								if (count($minmax[0]) == 2) {
+								// min und max
+									// min() von aktuellem Wert und erstem Wert der letzten Zeile, max() von aktuellm Wert und zweitem Wert der letzten Zeile
+									$write[$ise_id] = array();
+									foreach (( is_array($values[$ise_id]) ? $combine[$ise_id] : array($ise_id) ) as $key => $m_ise_id) {
+										$write[$ise_id][] = min(floatval($values[$m_ise_id]), $last[($key * 2)]);
+										$write[$ise_id][] = max(floatval($values[$m_ise_id]), $last[(($key * 2) + 1)]);
+										if (!empty($_verbose)) echo 'v  '.$ise_id.' '.$collect.' '.$m_ise_id.' letzte Werte '.$last[($key * 2)].' '.$last[(($key * 2) + 1)].', aktueller Wert '.$values[$m_ise_id].PHP_EOL;
+									}
+									$write[$ise_id] = precision($write[$ise_id], $arr['precision']);
+									$write[$ise_id] = implode(';', $write[$ise_id]);
+									if ($write[$ise_id] == implode(';', $last)) {
+										echo '- überspringe '.$collect.' '.$ise_id.' '.$history.', Werte sind unverändert '.str_replace(';', ' / ', $write[$ise_id]).PHP_EOL;
+										continue;
+									}
+									array_pop($csv);
+									echo '- '.$collect.'-Werte '.$ise_id.' '.$history.' alt '.implode(' / ', $last).' neu '.str_replace(';', ' / ', $write[$ise_id]).PHP_EOL;
+								} else {
+								// nur min oder max
+									// min() bzw. max() von aktuellen Werten und Werten der letzten Zeile
+									$m_value = array();
+									foreach (( is_array($values[$ise_id]) ? $combine[$ise_id] : array($ise_id) ) as $key => $m_ise_id) {
+										$m_value[] = precision(( preg_match('/min/i', $collect) ? min(floatval($values[$m_ise_id]), $last[$key]) : max(floatval($values[$m_ise_id]), $last[$key]) ), $arr['precision']);
+										if (!empty($_verbose)) echo 'v  '.$ise_id.' '.$collect.' '.$m_ise_id.' letzter Wert '.$last[$key].', aktueller Wert '.$values[$m_ise_id].PHP_EOL;
+									}
+									$write[$ise_id] = implode(';', $m_value);
+									if ($write[$ise_id] == implode(';', $last)) {
+										echo '- überspringe '.$collect.' '.$ise_id.' '.$history.', Wert ist unverändert '.str_replace(';', ' / ', $write[$ise_id]).PHP_EOL;
+										continue;
+									}
+									array_pop($csv);
+									echo '- '.$collect.'-Wert '.$ise_id.' '.$history.' alt '.implode(' / ', $last).' neu '.str_replace(';', ' / ', $write[$ise_id]).PHP_EOL;
+								}
+							} else {
+							// Neuer Tag, neue Zeile mit Werten anhängen. Für min+max aktuellen Wert als Min und Max aneinanderreihen.
+								if (count($minmax[0]) == 2) {
+									$m_value = '';
+									foreach (( is_array($values[$ise_id]) ? $combine[$ise_id] : array($ise_id) ) as $m_ise_id) {
+										$m_value .= str_repeat(precision(floatval($values[$m_ise_id]), $arr['precision']).';', 2);
+									}
+									$write[$ise_id] = rtrim($m_value, ';');
+								}
+								echo '- '.$collect.' '.$ise_id.' '.$history.', beginne neuen Tag mit '.str_replace(';', ' / ', $write[$ise_id]).PHP_EOL;
+							}
+						}
+					}
+
+					// Letzten gespeicherten Wert auslesen, überspringen falls unverändert
+					elseif (!empty($arr['only_changed'])) {
+						if (count($csv)) {
+
 							$last = explode(';', rtrim($csv[count($csv)-1], ';'), 2);
-							if ($last[1] == $values[$ise_id]) {
-								echo '- überspringe '.$ise_id.' '.$collect.' '.$history.', Wert ist unverändert '.$values[$ise_id].'.'.PHP_EOL;
+							if ($last[1] == $write[$ise_id]) {
+								echo '- überspringe '.$ise_id.' '.$collect.' '.$history.', Wert ist unverändert '.$write[$ise_id].'.'.PHP_EOL;
 								continue;
 							}
+
 						}
 					}
 
@@ -190,11 +306,13 @@ foreach ($diagramm as $ise_id => $collects) {
 				// cache Datei nicht vorhanden, anlegen und Schreirechte setzen
 
 					echo '- '.basename($cfile).' nicht vorhanden, lege neu an'.PHP_EOL;
-					touch($cfile);
-					if (!chmod($cfile, 0666)) echo 'Fehler beim setzen der Schreibrechte für '.$cfile.PHP_EOL;
+					if (empty($_dryrun)) {
+						touch($cfile);
+						if (!chmod($cfile, 0666)) echo 'Fehler beim setzen der Schreibrechte für '.$cfile.PHP_EOL;
+					}
 
 					// Prüfen, ob es eine cache Datei in altem Dateinamenformat (diagramm_<ise_id>_<history>.csv) gibt
-					$oldfile = __DIR__.'/cache/diagramm_'.preg_replace('/\D/', '-', $ise_id).'_'.$history.'.csv';
+					$oldfile = __DIR__.'/cache/diagramm_'.$ise_id.'_'.$history.'.csv';
 					if (file_exists($oldfile)) {
 					// alte Datei vorhanden, Inhalt einlesen und Datei löschen
 						echo '- Verschiebe Daten von '.basename($oldfile).' nach '.basename($cfile).PHP_EOL;
@@ -204,16 +322,21 @@ foreach ($diagramm as $ise_id => $collects) {
 
 					// keine Daten, leeres Array erzeugen
 					else $csv = array();
+
+					// Wert doppeln bei min+max, damit die neue Datei zwei Werte hat
+					if (preg_match('/min/i', $collect) and preg_match('/max/i', $collect)) {
+						$m_value = '';
+						foreach (( !empty($combine[$ise_id]) ? $combine[$ise_id] : array($ise_id) ) as $m_ise_id) {
+							$m_value .= str_repeat(precision(floatval($values[$m_ise_id]), $arr['precision']).';', 2);
+						}
+						$write[$ise_id] = rtrim($m_value, ';');
+					}
+
 				}
 
-				// Prefix für Beschriftung X-Achse
-				if (preg_match('/^(\d+):(\d+)$/', $collect)) $prefix = $tage[date('w')].' '.date('d.m.');
-				else $prefix = $tage[date('w')].' '.date('H:i');
-				# todo: Tag nur schreiben, wenn neuer Tag seit letztem Wert. Sollte mit $last[0] machbar sein #
-
-				$csv[] = $prefix.';'.$values[$ise_id].';';
-				echo '- schreibe '.$prefix.' = '.$values[$ise_id].' nach '.basename($cfile).PHP_EOL;
-				file_put_contents($cfile, implode("\n", $csv));
+				$csv[] = $prefix.';'.$write[$ise_id].';';
+				echo '- schreibe '.$prefix.' = '.$write[$ise_id].' nach '.basename($cfile).PHP_EOL;
+				if (empty($_dryrun)) file_put_contents($cfile, implode("\n", $csv));
 
 			}
 		}
